+++ tests/testthat/test-review_package.R
@@ -2,7 +2,7 @@ test_that("invalid package name with vector", {
   expect_true(
     grepl(
       "Invalid package name",
-      assert_package(name = letters, url = "xy"),
+      review_package(name = letters, url = "xy"),
       fixed = TRUE
     )
   )
@@ -12,7 +12,7 @@ test_that("invalid package name with dot", {
   expect_true(
     grepl(
       "Invalid package name",
-      assert_package(
+      review_package(
         name = ".gh",
         url = "https://github.com/r-lib/gh"
       ),
@@ -25,7 +25,7 @@ test_that("custom JSON", {
   expect_true(
     grepl(
       "looks like JSON",
-      assert_package(name = "xy", url = "{"),
+      review_package(name = "xy", url = "{"),
       fixed = TRUE
     )
   )
@@ -35,7 +35,7 @@ test_that("advisory", {
   expect_true(
     grepl(
       "advisory",
-      assert_package(
+      review_package(
         name = "def",
         url = "https://github.com/abc/def",
         advisories = c("abc", "def")
@@ -49,7 +49,7 @@ test_that("invalid vector URL", {
   expect_true(
     grepl(
       "Invalid package URL",
-      assert_package(
+      review_package(
         name = "xy",
         url = letters
       ),
@@ -62,7 +62,7 @@ test_that("malformed URL", {
   expect_true(
     grepl(
       "Found malformed URL",
-      assert_package(
+      review_package(
         name = "gh",
         url = "github.com/r-lib/gh"
       ),
@@ -75,7 +75,7 @@ test_that("package name/repo disagreement", {
   expect_true(
     grepl(
       "appears to disagree with the repository name in the URL",
-      assert_package(
+      review_package(
         name = "gh2",
         url = "https://github.com/r-lib/gh"
       ),
@@ -88,7 +88,7 @@ test_that("https", {
   expect_true(
     grepl(
       "is not https",
-      assert_package(
+      review_package(
         name = "gh",
         url = "http://github.com/r-lib/gh"
       ),
@@ -101,9 +101,10 @@ test_that("GitHub/GitLab URL", {
   expect_true(
     grepl(
       "is not a GitHub or GitLab URL",
-      assert_package(
+      review_package(
         name = "gh",
-        url = "https://github.gov/r-lib/gh"
+        url = "https://github.gov/r-lib/gh",
+        advisories = character(0L)
       ),
       fixed = TRUE
     )
@@ -114,9 +115,10 @@ test_that("owner URL", {
   expect_true(
     grepl(
       "appears to be an owner",
-      assert_package(
+      review_package(
         name = "gh",
-        url = "https://github.com/gh"
+        url = "https://github.com/gh",
+        advisories = character(0L)
       ),
       fixed = TRUE
     )
@@ -127,9 +129,10 @@ test_that("CRAN mirror", {
   expect_true(
     grepl(
       "appears to use a CRAN mirror",
-      assert_package(
+      review_package(
         name = "gh",
-        url = "https://github.com/cran/gh"
+        url = "https://github.com/cran/gh",
+        advisories = character(0L)
       ),
       fixed = TRUE
     )
@@ -153,7 +156,7 @@ test_that("HTTP error", {
   expect_true(
     grepl(
       "returned HTTP error",
-      assert_package(
+      review_package(
         name = "afantasticallylongandimpossiblepackage",
         url = "https://github.com/r-lib/afantasticallylongandimpossiblepackage"
       ),
@@ -166,7 +169,7 @@ test_that("release URL", {
   expect_true(
     grepl(
       "No full release found at URL",
-      assert_package(
+      review_package(
         name = "test.no.release",
         url = "https://github.com/wlandau/test.no.release"
       ),
@@ -177,9 +180,10 @@ test_that("release URL", {
 
 test_that("good GitHub registration", {
   expect_null(
-    assert_package(
+    review_package(
       name = "gh",
-      url = "https://github.com/r-lib/gh"
+      url = "https://github.com/r-lib/gh",
+      advisories = character(0L)
     )
   )
 })
@@ -187,7 +191,7 @@ test_that("good GitHub registration", {
 test_that("GitLab registration with free-form license", {
   tmp <- utils::capture.output(
     suppressMessages(
-      out <- assert_package(
+      out <- review_package(
         name = "test",
         url = "https://gitlab.com/wlandau/test"
       )